refactor(api): Split out protocol_api.__init__ to submodules
This file was getting a bit out of hand and promising to get much worse.

The context classes are now in contexts and the deck is in geometry. Other
geometry-related, context-free utilities should be placed there as well.
